Made execution policy configurable

+++ source/fern/python_extension/algorithm/algebra/elementary/test/add_test.py
@@ -71,7 +71,7 @@ class AddTest(ft.TestCase):
             [  6,            7]]
         mask_we_want = self.mask
         raster_we_want = self.masked_raster(values_we_want, mask_we_want,
-            value_type=fern.float64)
+            value_type=fern.int64)
 
         self.assertMaskedRasterEqual(raster_we_got, raster_we_want)
 
@@ -86,6 +86,6 @@ class AddTest(ft.TestCase):
             [  6,            7]]
         mask_we_want = self.mask
         raster_we_want = self.masked_raster(values_we_want, mask_we_want,
-            value_type=fern.float64)
+            value_type=fern.int64)
 
         self.assertMaskedRasterEqual(raster_we_got, raster_we_want)
